requirements.txt
pyscreenshot
tensorflow
scikit-learn
matplotlib
selenium
pillow
bs4